fix: move tryParseAmount to shared package
