+++ monorepo/packages/shared/src/utils/index.tsx
@@ -4,6 +4,7 @@ import listVersionLabel from 'src/utils/listVersionLabel';
 import { maxAmountSpend } from 'src/utils/maxAmountSpend';
 import { parseENSAddress } from 'src/utils/parseENSAddress';
 import uriToHttp from 'src/utils/uriToHttp';
+export { tryParseAmount } from 'src/utils/tryParseAmount';
 
 export { chunkArray, isZero, uriToHttp, parseENSAddress, maxAmountSpend, listVersionLabel };
 export * from './common';
