+++ monorepo/packages/swap/src/state/pswap/hooks/common.ts
@@ -1,5 +1,4 @@
 /* eslint-disable max-lines */
-import { parseUnits } from '@ethersproject/units';
 import { Order, useGelatoLimitOrdersHistory, useGelatoLimitOrdersLib } from '@gelatonetwork/limit-orders-react';
 import {
   CAVAX,
@@ -8,7 +7,6 @@ import {
   CurrencyAmount,
   ElixirTrade,
   FACTORY_ADDRESS,
-  JSBI,
   Price,
   Token,
   TokenAmount,
@@ -101,30 +99,6 @@ export function useSwapActionHandlers(chainId: ChainId): {
     onUserInput,
     onChangeRecipient,
   };
-}
-
-// try to parse a user entered amount for a given token
-export function tryParseAmount(
-  value?: string,
-  currency?: Currency,
-  chainId: ChainId = ChainId.AVALANCHE,
-): CurrencyAmount | undefined {
-  if (!value || !currency) {
-    return undefined;
-  }
-  try {
-    const typedValueParsed = parseUnits(value, currency.decimals).toString();
-    if (typedValueParsed !== '0') {
-      return currency instanceof Token
-        ? new TokenAmount(currency, JSBI.BigInt(typedValueParsed))
-        : CurrencyAmount.ether(JSBI.BigInt(typedValueParsed), chainId);
-    }
-  } catch (error) {
-    // should fail if the user specifies too many decimal places of precision (or maybe exceed max uint?)
-    console.debug(`Failed to parse input amount: "${value}"`, error);
-  }
-  // necessary for all paths to return a value
-  return undefined;
 }
 
 const BAD_RECIPIENT_ADDRESSES: string[] = [
